+++ neoforge/src/main/java/io/github/mortuusars/envelope/neoforge/event/NeoForgeCommonEvents.java
@@ -110,7 +110,7 @@ public class NeoForgeCommonEvents {
     public static void registerSpawnPlacements(RegisterSpawnPlacementsEvent event) {
         event.register(Envelope.EntityTypes.PIGEON.get(), SpawnPlacementTypes.ON_GROUND, Heightmap.Types.MOTION_BLOCKING,
               Pigeon::checkSpawnRules, RegisterSpawnPlacementsEvent.Operation.REPLACE);
-        event.register(Envelope.EntityTypes.CHARRED_PIGEON.get(), SpawnPlacementTypes.NO_RESTRICTIONS, Heightmap.Types.MOTION_BLOCKING,
+        event.register(Envelope.EntityTypes.CHARRED_PIGEON.get(), SpawnPlacementTypes.ON_GROUND, Heightmap.Types.MOTION_BLOCKING,
               CharredPigeon::checkSpawnRules, RegisterSpawnPlacementsEvent.Operation.REPLACE);
     }